Write doc for shouldOpenGroupByDefault
