+++ readme.md
@@ -291,6 +291,32 @@ reportFileSizeImpact({
 </details>
 
 <details>
+  <summary>shouldOpenGroupByDefault parameter</summary>
+
+`shouldOpenGroupByDefault` parameter is a function received named arguments and returning a boolean. When the returned boolean is true, the group is opened by default in the pull request comment.
+
+The following code would always open `critical files` group.
+
+```js
+import { reportFileSizeImpact } from "@jsenv/file-size-impact"
+
+reportFileSizeImpact({
+  trackingConfig: {
+    "critical files": {
+      "./dist/main.js": true,
+    },
+    "remaining files": {
+      "./dist/**/*.js": true,
+      "./dist/main.js": false,
+    },
+  },
+  shouldOpenGroupByDefault: ({ groupName }) => groupName === "critical files",
+})
+```
+
+</details>
+
+<details>
   <summary>transformations parameter</summary>
 
 `transformations` parameter is an object used to transform files content before computing their size. This parameter is optional with a default tracking file size without transformation called `raw`.
